add daily agenda view
- also rename EventListing to Event
